refactor and increase fontsize for curated search thumbnail images
